server/requirements.txt
cryptography==37.0.2
Flask==2.2.2
Flask_Cors==3.0.10
Flask-Login==0.6.3
Flask_Migrate==4.0.0
flask_sqlalchemy==3.0.2
itsdangerous==2.1.2
matplotlib==3.6.2
networkx==2.8.8
numpy==1.23.5
pandas==1.5.2
pytest==7.2.0
PyMySQL==1.0.2 
setuptools==58.1.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
PyJWT==2.8.0